refactor: `IndexData` Entity와 `IndexDataRepository`를 각각 `entity` 패키지와 `repository` 패키지로 이동
